clean code

+++ app/src/main/java/au/com/swivelgroup/newsapp/data/model/Status.kt
@@ -3,6 +3,6 @@ package au.com.swivelgroup.newsapp.data.model
 import au.com.swivelgroup.newsapp.base.BaseCallBack
 
 data class Status(
-    @BaseCallBack.Companion.ServerCode val code: Long, val status: String,
+    @BaseCallBack.Companion.ServerCode val code: Int, val status: String,
     val message: String
 )
